pneumoRL/requirements.txt
backcall==0.1.0
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
kaggle==1.4.7.1
line-profiler==2.1.2
nibabel==2.3.0
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
pydicom==1.1.0
Pygments==2.2.0
python-dateutil==2.7.3
python-slugify==1.2.6
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
torch==0.4.1
torchsample==0.1.3
torchvision==0.2.1
tqdm==4.26.0
traitlets==4.3.2
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
